Changing the way how page is generated

+++ cmd/gui/mvc/model/status.go
@@ -5,23 +5,22 @@ import (
 )
 
 // System Ststus
-type
-	DuoUIstatus struct {
-		Version    string `json:"ver"`
-		UpTime     int64  `json:"uptime"`
-		CurrentNet string `json:"net"`
-		Chain      string `json:"chain"`
-		Node       *NodeStatus
-		Wallet     *WalletStatus
-		Kopach     *KopachStatus
-	}
+type DuoUIstatus struct {
+	Version    string `json:"ver"`
+	UpTime     int64  `json:"uptime"`
+	CurrentNet string `json:"net"`
+	Chain      string `json:"chain"`
+	Node       *NodeStatus
+	Wallet     *WalletStatus
+	Kopach     *KopachStatus
+}
 
 type NodeStatus struct {
 	NetHash          int64
-	BlockHeight      int32
+	BlockHeight      int
 	BestBlock        string
 	Difficulty       float64
-	BlockCount       int64
+	BlockCount       int
 	NetworkLastBlock int32
 	ConnectionCount  int32
 }
@@ -37,29 +36,20 @@ type WalletStatus struct {
 	Txs           *DuoUItransactionsExcerpts
 	LastTxs       *DuoUItransactions
 }
-type
-	DuoUIhashes struct{ int64 }
-type
-	DuoUInetworkHash struct{ int64 }
-type
-	DuoUIheight struct{ int32 }
-type
-	DuoUIbestBlockHash struct{ string }
-type
-	DuoUIdifficulty struct{ float64 }
+type DuoUIhashes struct{ int64 }
+type DuoUInetworkHash struct{ int64 }
+type DuoUIheight struct{ int32 }
+type DuoUIbestBlockHash struct{ string }
+type DuoUIdifficulty struct{ float64 }
 
 //type
 // MempoolInfo      struct { string}
-type
-	DuoUIblockCount struct{ int64 }
-type
-	DuoUInetLastBlock struct{ int32 }
-type
-	DuoUIconnections struct{ int32 }
-type
-	DuoUIlocalHost struct {
-		//Cpu        []cpu.InfoStat        `json:"cpu"`
-		//CpuPercent []float64             `json:"cpupercent"`
-		//Memory     mem.VirtualMemoryStat `json:"mem"`
-		//Disk       disk.UsageStat        `json:"disk"`
-	}
+type DuoUIblockCount struct{ int64 }
+type DuoUInetLastBlock struct{ int32 }
+type DuoUIconnections struct{ int32 }
+type DuoUIlocalHost struct {
+	//Cpu        []cpu.InfoStat        `json:"cpu"`
+	//CpuPercent []float64             `json:"cpupercent"`
+	//Memory     mem.VirtualMemoryStat `json:"mem"`
+	//Disk       disk.UsageStat        `json:"disk"`
+}
